blog.php
git-svn-id: https://svn.liip.ch/repos/public/fluxcms/trunk@9110 99f7f646-bdd5-0310-89f9-ddcb10b10bf3

+++ inc/bx/plugins/blog.php
@@ -173,7 +173,7 @@ class bx_plugins_blog extends bx_plugin implements bxIplugin {
                 return $xml;
             }
         }
-        $startEntry  = isset($_GET['start']) ? $_GET['start'] : 0;
+        $startEntry  = (int) isset($_GET['start']) ? $_GET['start'] : 0;
         if (($pos = strrpos($id,"/")) > 0) {
             $cat = substr($id,0,$pos);
             $id = substr($id, $pos + 1);
